Update Migration Student, Student Old

+++ backend/database/migrations/2025_07_09_131651_create_students_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('family_rank', 50);
             $table->enum('citizenship', ['Indonesia', 'Non Indonesia']);
             $table->string('country', 100)->nullable();
-            $table->unsignedTinyInteger('nik')->nullable(); 
+            $table->unsignedBigInteger('nik')->nullable(); 
             $table->string('kitas', 50)->nullable();
             $table->string('place_of_birth', 100);
             $table->date('date_of_birth');
